// src/store/user/types.ts
export const enum UserActionTypes {
	LOGIN_USER = 'LOGIN_USER',
	LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS',
	LOGIN_USER_ERROR = 'LOGIN_USER_ERROR',
	SET_ROLE = 'SET_ROLE',

	REGISTER_USER = 'REGISTER_USER',
	REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS',
	REGISTER_USER_ERROR = 'REGISTER_USER_ERROR',

	LOGOUT_USER = 'LOGOUT_USER',
	LOGOUT_USER_SUCCESS = 'LOGOUT_USER_SUCCESS',

	GET_USER = 'GET_USER',
	GET_USER_SUCCESS = 'GET_USER_SUCCESS',
	GET_USER_ERROR = 'GET_USER_ERROR',

	GET_SPECIALIZATIONS = 'GET_SPECIALIZATIONS',
	GET_SPECIALIZATIONS_SUCCESS = 'GET_SPECIALIZATIONS_SUCCESS',

	GET_MY_STUDENTS = 'GET_MY_STUDENTS',
	GET_MY_STUDENTS_SUCCESS = 'GET_MY_STUDENTS_SUCCESS',

	GET_MY_TRAINERS = 'GET_MY_TRAINERS',
	GET_MY_TRAINERS_SUCCESS = 'GET_MY_TRAINERS_SUCCESS',

	CHANGE_PASSWORD = 'CHANGE_PASSWORD',
	CHANGE_PASSWORD_SUCCESS = 'CHANGE_PASSWORD_SUCCESS',
	CHANGE_PASSWORD_ERROR = 'CHANGE_PASSWORD_ERROR',

	GET_ALL_TRAINERS = 'GET_ALL_TRAINERS',
	GET_ALL_TRAINERS_SUCCESS = 'GET_ALL_TRAINERS_SUCCESS',
	GET_ALL_TRAINERS_ERROR = 'GET_ALL_TRAINERS_ERROR',

	ADD_TRAINERS = 'ADD_TRAINERS',
	ADD_TRAINERS_SUCCESS = 'ADD_TRAINERS_SUCCESS',
	ADD_TRAINERS_ERROR = 'ADD_TRAINERS_ERROR',

	EDIT_USER = 'EDIT_USER',
	EDIT_USER_SUCCESS = 'EDIT_USER_SUCCESS',
	EDIT_USER_ERROR = 'EDIT_USER_ERROR',

	DELETE_USER = 'DELETE_USER',
	DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS',
	DELETE_USER_ERROR = 'DELETE_USER_ERROR',
}